+++ DDL/follow_users.sql
@@ -1,7 +1,7 @@
 CREATE TABLE `follow_users` (
-  `id` int NOT NULL AUTO_INCREMENT,
-  `from` int NOT NULL,
-  `to` int NOT NULL,
+  `id` bigint NOT NULL AUTO_INCREMENT,
+  `from` bigint NOT NULL,
+  `to` bigint NOT NULL,
   PRIMARY KEY (`id`)
 ) ENGINE = InnoDB DEFAULT CHARSET = utf8;
