+++ DDL/shops.sql
@@ -1,8 +1,8 @@
 CREATE TABLE `shops` (
-  `id` int NOT NULL AUTO_INCREMENT,
+  `id` bigint NOT NULL AUTO_INCREMENT,
   `name` varchar(100) NOT NULL,
-  `station_id` int DEFAULT 0,
-  `create_user` int NOT NULL DEFAULT 1,
+  `station_id` bigint DEFAULT 0,
+  `create_user` bigint NOT NULL DEFAULT 1,
   `is_deleted` boolean DEFAULT 0,
   `create_at` datetime NOT NULL,
   `delete_at` datetime,
